+++ ZenUnitTests/Utils/Concurrency/Mock/FuturistMock.h
@@ -6,5 +6,5 @@ struct FuturistMock : public Zen::Mock<Futurist<ThisPointerType>>
 {
    using VoidMemberFunctionType = void(ThisPointerType::*)();
    ZENMOCK_NONVOID2_CONST(shared_ptr<VoidFuture>, Async, VoidMemberFunctionType, ThisPointerType*)
-   ZENMOCK_NONVOID2_CONST(future_status, WaitMilliseconds, future<void>*, long long)
+   ZENMOCK_NONVOID2_CONST(future_status, WaitMilliseconds, future<void>*, unsigned)
 };
